Consolidate guest image generator support
Move guest-context singleton creation and registration helpers out of NativeImageGenerator. Keep
the builder orchestration focused while collecting guest object construction and proxy wiring in
one package-private support class without changing behavior.
