+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/NativeImageGenerator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2012, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2012, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -33,9 +33,7 @@ import static jdk.graal.compiler.replacements.StandardGraphBuilderPlugins.regist
 
 import java.io.IOException;
 import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
-import java.nio.ByteOrder;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -175,10 +173,7 @@ import com.oracle.svm.core.util.ExitStatus;
 import com.oracle.svm.core.util.InterruptImageBuilding;
 import com.oracle.svm.core.util.UserError;
 import com.oracle.svm.core.util.UserErrorSupportImpl;
-import com.oracle.svm.guest.staging.ArgsSupport;
-import com.oracle.svm.guest.staging.JavaMainSupport;
 import com.oracle.svm.guest.staging.config.SubstrateGuestLibC;
-import com.oracle.svm.guest.staging.config.SubstrateGuestTarget;
 import com.oracle.svm.guest.staging.jdk.RuntimeSupport;
 import com.oracle.svm.guest.staging.option.RuntimeOptionValidationSupport;
 import com.oracle.svm.guest.staging.option.RuntimeOptionValues;
@@ -342,7 +337,6 @@ import jdk.graal.compiler.phases.util.Providers;
 import jdk.graal.compiler.printer.GraalDebugHandlersFactory;
 import jdk.graal.compiler.replacements.NodeIntrinsificationProvider;
 import jdk.graal.compiler.replacements.TargetGraphBuilderPlugins;
-import jdk.graal.compiler.vmaccess.InvocationException;
 import jdk.graal.compiler.word.WordOperationPlugin;
 import jdk.graal.compiler.word.WordTypes;
 import jdk.internal.loader.ClassLoaders;
@@ -362,8 +356,6 @@ import jdk.vm.ci.meta.ResolvedJavaType;
 import jdk.vm.ci.riscv64.RISCV64;
 
 public class NativeImageGenerator {
-    private static final String WINDOWS_ARGS_SUPPORT_CLASS_NAME = "com.oracle.svm.core.windows.WindowsJavaMainWrapperArgsSupport";
-
     protected final FeatureHandler featureHandler;
     protected final ImageClassLoader loader;
     protected final HostedOptionProvider optionProvider;
@@ -1042,12 +1034,12 @@ public class NativeImageGenerator {
                 SubstrateTarget target = createTarget();
                 ImageSingletons.add(Platform.class, loader.platform);
                 ImageSingletons.add(SubstrateTarget.class, target);
-                setupGuestTargetDescription(target);
+                GuestImageGeneratorSupport.setupTargetDescription(target);
 
                 ImageSingletons.add(SubstrateOptions.ReportingSupport.class, new SubstrateOptions.ReportingSupport(
                                 DiagnosticsMode.getValue() ? DiagnosticsDir.getValue().lastValue().get() : Path.of("reports")));
                 FutureDefaultsOptions.parseAndVerifyOptions();
-                installArgsSupport();
+                GuestImageGeneratorSupport.installArgsSupport();
                 if (javaMainMethod != null) {
                     installJavaMainSupport(javaMainMethod);
                 }
@@ -1255,89 +1247,10 @@ public class NativeImageGenerator {
     }
 
     /**
-     * Installs the guest-staging argument support singleton after the image singleton registries
-     * have been installed.
-     * <p>
-     * {@link ArgsSupport} is used by runtime code in guest staging, so fully isolated builds need
-     * the singleton instance in the guest registry. The implementation object is therefore created
-     * in the guest context and registered through {@link GuestImageSingletonSupport}. The singleton
-     * keeps the {@code InitialLayerOnly} layer contract by following the same loaded-key check as
-     * automatic singleton registration: if layer loading already handled the key, setup does not
-     * create or register a replacement object. GR-76716 tracks direct guest-staging support for
-     * automatic singleton registration, which would let this become an ordinary guest-staging
-     * registration. GR-76886 tracks moving the Windows-specific implementation into guest staging
-     * once the required platform bindings and platform-specific registration selection are
-     * guest-staging owned.
-     */
-    private static void installArgsSupport() {
-        if (ImageSingletons.lookup(LoadedLayeredImageSingletonInfo.class).handledDuringLoading(ArgsSupport.class)) {
-            return;
-        }
-
-        GuestAccess access = GuestAccess.get();
-        ResolvedJavaType argsSupportType = access.lookupType(ArgsSupport.class);
-        ResolvedJavaType implementationType = getArgsSupportImplementationType(access);
-        ResolvedJavaMethod ctor = JVMCIReflectionUtil.getDeclaredConstructor(implementationType);
-        JavaConstant argsSupport;
-        try {
-            argsSupport = access.invoke(ctor, null);
-        } catch (InvocationException ex) {
-            throw VMError.shouldNotReachHere("Error creating Java argument support in the guest context", ex);
-        }
-        GuestImageSingletonSupport.add(argsSupportType, argsSupport);
-    }
-
-    /**
-     * Returns the guest type that should implement the {@link ArgsSupport} singleton for the target
-     * platform.
-     */
-    private static ResolvedJavaType getArgsSupportImplementationType(GuestAccess access) {
-        if (Platform.includedIn(Platform.WINDOWS.class)) {
-            return access.lookupType(WINDOWS_ARGS_SUPPORT_CLASS_NAME);
-        }
-        return access.lookupType(ArgsSupport.class);
-    }
-
-    /**
-     * Installs Java-main state after the image singleton registries have been installed.
-     * <p>
-     * {@link NativeImageGeneratorRunner} resolves the application Java main method. This method runs
-     * in the generator setup phase that has active builder and guest singleton registries. The
-     * {@link JavaMainSupport} object is constructed in the guest context because it owns method
-     * handles for guest methods, then registered through {@link GuestImageSingletonSupport} so the
-     * builder-side code path performs all guest singleton registration consistently.
-     *
-     * @param javaMainMethod the application Java main method resolved by
-     *            {@link NativeImageGeneratorRunner}
+     * Installs the Java-main support selected by this image generator.
      */
     protected void installJavaMainSupport(ResolvedJavaMethod javaMainMethod) {
-        GuestAccess access = GuestAccess.get();
-        JavaConstant executable = access.asExecutableConstant(javaMainMethod);
-        if (executable == null) {
-            throw UserError.abort("Cannot install Java main support because no reflective executable is available for %s.", javaMainMethod.format("%H.%n(%p)"));
-        }
-
-        ResolvedJavaType javaMainSupportType = access.lookupType(JavaMainSupport.class);
-        ResolvedJavaMethod ctor = JVMCIReflectionUtil.getDeclaredConstructor(access.getProviders().getMetaAccess(), javaMainSupportType, Method.class);
-        JavaConstant javaMainSupport;
-        try {
-            javaMainSupport = access.invoke(ctor, null, executable);
-        } catch (InvocationException ex) {
-            if (ex.getCause() instanceof IllegalArgumentException iae) {
-                throw UserError.abort(iae, "%s", iae.getMessage());
-            }
-            throw VMError.shouldNotReachHere("Error creating Java main support in the guest context", ex);
-        }
-        GuestImageSingletonSupport.add(javaMainSupportType, javaMainSupport);
-    }
-
-    private static JavaConstant fromEnum(Enum<?> kind) {
-        GuestAccess access = GuestAccess.get();
-        ResolvedJavaType enumType = access.getProviders().getMetaAccess().lookupJavaType(kind.getDeclaringClass());
-        JavaConstant enumName = access.asGuestString(kind.name());
-        ResolvedJavaMethod valueOf = JVMCIReflectionUtil.getUniqueDeclaredMethod(access.getProviders().getMetaAccess(), enumType, "valueOf", String.class);
-        JavaKind.valueOf(kind.name());
-        return access.invoke(valueOf, null, enumName);
+        GuestImageGeneratorSupport.installJavaMainSupport(javaMainMethod);
     }
 
     /**
@@ -1355,28 +1268,8 @@ public class NativeImageGenerator {
         NativeImageGenerator.loadAndInstallLayeredSingletons(imageLayerSupport, hostedSingletonManagement);
         if (GuestAccess.get().isFullyIsolated()) {
             /* Install a second singleton registry in the guest context. */
-            GuestImageSingletonSupport.install();
-            registerGuestImageLayerBuildingSupport(imageLayerSupport);
+            GuestImageGeneratorSupport.installSingletonRegistry(imageLayerSupport);
         }
-    }
-
-    private static void registerGuestImageLayerBuildingSupport(HostedImageLayerBuildingSupport imageLayerSupport) {
-        GuestAccess access = GuestAccess.get();
-        ResolvedJavaType key = access.lookupType(ImageLayerBuildingSupportProvider.class);
-        JavaConstant hostProxy = access.createHostProxy(imageLayerSupport, key);
-        GuestImageSingletonSupport.add(key, hostProxy);
-    }
-
-    private static void setupGuestTargetDescription(SubstrateTarget target) {
-        GuestAccess access = GuestAccess.get();
-        ResolvedJavaMethod ctor = JVMCIReflectionUtil.getDeclaredConstructor(access.getProviders().getMetaAccess(), SubstrateGuestTarget.class, JavaKind.class, int.class, ByteOrder.class);
-
-        JavaConstant wordKind = fromEnum(target.wordJavaKind);
-        JavaConstant wordSize = JavaConstant.forInt(target.wordSize);
-        JavaConstant byteOrder = JVMCIReflectionUtil.readStaticField(access.elements.java_nio_ByteOrder, target.arch.getByteOrder().toString());
-        JavaConstant guestTargetDescription = access.invoke(ctor, null, wordKind, wordSize, byteOrder);
-
-        GuestImageSingletonSupport.add(SubstrateGuestTarget.class, guestTargetDescription);
     }
 
     /**
